Make trace module *mostly* work with python3
This is enough compatibility tweaks to use python3 without
crashing (tested with some outside app I've been working on).

A quick regression test with python2.7 was promising,
showed consistent trace output before and after these changes.
However, the python3 output is not quite right yet: the class names get
dropped when tracing normal methods.

This has to do with python3 doing away with the distinction between
class-less functions and unbound methods. I stopped a crash on undefined
types.UnboundMethodType by instead referring to the detected type of a
sample class method (like was already being done for some special method
types) but on python3 this detected type will be the same as for a
normal function.  So it's no wonder class methods are being mistaken for
normal functions, and their class fails to be listed in the trace.

Other changes for python3 compatibility:
* thread-local variables switched from 'thread' to 'threading' module
* switched to new naming scheme for function internals,
  since luckily python2.7 supports it now, e.g. func_code -> __code__
* check against tuple of possible string types (bytes, str) instead
  of checking for an instance of basestring which went away in python3