Both static and class methods work on python3
The hack from bdcfb941888 did not actually work so well,
so I switched to a different strategy for deciding which
type of function is being wrapped -- pass the function type
in from the caller, where that information is available
from the class object, rather than trying to deduce it based
on the function object for the method.

This new strategy is so much cleaner that the same code actually
works for both python2 and python3 -- good bye to some ugly
hacks based on try-except!

Also simplified the function wrapping while working on grokking
the code better to find the new approach:

* __rewrap was classmethod() for class methods and a no-op the
  rest of the time, so now I'm just wrapping with classmethod()
  in the spot where a class-method wrapper has just been returned.

* creating a partial object with wraps() seems silly when
  immediately passing in the other parameter; skip the partial
  and call update_wrapper() directly